refactor(peagen): organize gateway rpc handlers
