+++ pkgs/standards/peagen/peagen/gateway/__init__.py
@@ -25,7 +25,6 @@ from fastapi import FastAPI, Request, Response, HTTPException
 from peagen.plugins.queues import QueueBase
 
 from peagen.transport import RPCDispatcher, RPCRequest
-from peagen.transport.jsonrpc import RPCException
 from peagen.orm import Base, Task
 from peagen.orm.status import Status
 from peagen.schemas import TaskRead, TaskCreate, TaskUpdate
@@ -51,14 +50,20 @@ from peagen.errors import (
     MissingActionError,
     MigrationFailureError,
     NoWorkerAvailableError,
-    TaskNotFoundError,
 )
 import peagen.defaults as defaults
 from peagen.defaults import BAN_THRESHOLD
 from peagen.defaults.error_codes import ErrorCode
 from peagen.core import migrate_core
-from peagen.core.task_core import get_task_result
 from peagen.services import create_task, get_task, update_task
+
+from .rpc import (  # noqa: F401
+    keys as _rpc_keys,
+    pool as _rpc_pool,
+    secrets as _rpc_secrets,
+    tasks as _rpc_tasks,
+    workers as _rpc_workers,
+)
 
 _db = reload(_db)
 engine = _db.engine
@@ -535,384 +540,6 @@ async def rpc_endpoint(request: Request):
     return Response(
         content=json.dumps(resp), status_code=status, media_type="application/json"
     )
-
-
-# ─────────────────────────── Key/Secret RPC ─────────────────────
-
-
-@rpc.method("Keys.upload")
-async def keys_upload(public_key: str) -> dict:
-    """Store a trusted public key."""
-    key = pgpy.PGPKey()
-    key.parse(public_key)
-    TRUSTED_USERS[key.fingerprint] = public_key
-    log.info("key uploaded: %s", key.fingerprint)
-    return {"fingerprint": key.fingerprint}
-
-
-@rpc.method("Keys.fetch")
-async def keys_fetch() -> dict:
-    """Return all trusted keys indexed by fingerprint."""
-    return TRUSTED_USERS
-
-
-@rpc.method("Keys.delete")
-async def keys_delete(fingerprint: str) -> dict:
-    """Remove a public key by its fingerprint."""
-    TRUSTED_USERS.pop(fingerprint, None)
-    log.info("key removed: %s", fingerprint)
-    return {"ok": True}
-
-
-@rpc.method("Secrets.add")
-async def secrets_add(
-    name: str,
-    secret: str,
-    tenant_id: str = "default",
-    owner_fpr: str = "unknown",
-    version: int | None = None,
-) -> dict:
-    """Store an encrypted secret."""
-    async with Session() as session:
-        await upsert_secret(session, tenant_id, owner_fpr, name, secret)
-        await session.commit()
-    log.info("secret stored: %s", name)
-    return {"ok": True}
-
-
-@rpc.method("Secrets.get")
-async def secrets_get(name: str, tenant_id: str = "default") -> dict:
-    """Retrieve an encrypted secret."""
-    async with Session() as session:
-        row = await fetch_secret(session, tenant_id, name)
-    if not row:
-        raise RPCException(
-            code=ErrorCode.SECRET_NOT_FOUND,
-            message="secret not found",
-        )
-    return {"secret": row.cipher}
-
-
-@rpc.method("Secrets.delete")
-async def secrets_delete(
-    name: str,
-    tenant_id: str = "default",
-    version: int | None = None,
-) -> dict:
-    """Remove a secret by name."""
-    async with Session() as session:
-        await delete_secret(session, tenant_id, name)
-        await session.commit()
-    log.info("secret removed: %s", name)
-
-    return {"ok": True}
-
-
-# ─────────────────────────── Pool RPCs ──────────────────────────
-@rpc.method("Pool.create")
-async def pool_create(name: str):
-    await queue.sadd("pools", name)
-    log.info("pool created: %s", name)
-    return {"name": name}
-
-
-@rpc.method("Pool.join")
-async def pool_join(name: str):
-    member = str(uuid.uuid4())[:8]
-    await queue.sadd(f"pool:{name}:members", member)
-    log.info("member %s joined pool %s", member, name)
-    return {"memberId": member}
-
-
-# ─────────────────────────── Task RPCs ──────────────────────────
-@rpc.method("Task.submit")
-async def task_submit(
-    pool: str,
-    payload: dict,
-    taskId: str | None = None,
-    deps: list[str] | None = None,
-    edge_pred: str | None = None,
-    labels: list[str] | None = None,
-    in_degree: int | None = None,
-    config_toml: str | None = None,
-):
-    await queue.sadd("pools", pool)  # track pool even if not created
-
-    action = (payload or {}).get("action")
-    handlers: set[str] = set()
-    for w in await _live_workers_by_pool(pool):
-        raw = w.get("handlers", [])
-        if isinstance(raw, str):
-            try:
-                raw = json.loads(raw)
-            except Exception:  # noqa: BLE001
-                raw = []
-        handlers.update(raw)
-    if action is not None and action not in handlers:
-        raise RPCException(
-            code=-32601, message="Method not found", data={"method": str(action)}
-        )
-
-    if taskId and await _load_task(taskId):
-        new_id = str(uuid.uuid4())
-        log.warning("task id collision: %s → %s", taskId, new_id)
-        taskId = new_id
-
-    task = Task(
-        id=taskId or str(uuid.uuid4()),
-        pool=pool,
-        payload=payload,
-        deps=deps or [],
-        edge_pred=edge_pred,
-        labels=labels or [],
-        in_degree=in_degree or 0,
-        config_toml=config_toml,
-    )
-
-    # 1) put on the queue for the scheduler
-    await queue.rpush(f"{READY_QUEUE}:{pool}", task.model_dump_json())
-    await _publish_queue_length(pool)
-
-    # 2) save hash + TTL
-    await _save_task(task)
-
-    # 3) persist to Postgres & emit CloudEvent (helpers you already have)
-    await _persist(task)
-    await _publish_task(task)
-
-    log.info("task %s queued in %s (ttl=%ss)", task.id, pool, TASK_TTL)
-    return {"taskId": task.id}
-
-
-@rpc.method("Task.cancel")
-async def task_cancel(selector: str):
-    targets = await _select_tasks(selector)
-    from peagen.handlers import control_handler
-
-    count = await control_handler.apply("cancel", queue, targets, READY_QUEUE, TASK_TTL)
-    log.info("cancel %s -> %d tasks", selector, count)
-    return {"count": count}
-
-
-@rpc.method("Task.pause")
-async def task_pause(selector: str):
-    targets = await _select_tasks(selector)
-    from peagen.handlers import control_handler
-
-    count = await control_handler.apply("pause", queue, targets, READY_QUEUE, TASK_TTL)
-    log.info("pause %s -> %d tasks", selector, count)
-    return {"count": count}
-
-
-@rpc.method("Task.resume")
-async def task_resume(selector: str):
-    targets = await _select_tasks(selector)
-    from peagen.handlers import control_handler
-
-    count = await control_handler.apply("resume", queue, targets, READY_QUEUE, TASK_TTL)
-    log.info("resume %s -> %d tasks", selector, count)
-    return {"count": count}
-
-
-@rpc.method("Task.retry")
-async def task_retry(selector: str):
-    targets = await _select_tasks(selector)
-    from peagen.handlers import control_handler
-
-    count = await control_handler.apply("retry", queue, targets, READY_QUEUE, TASK_TTL)
-    log.info("retry %s -> %d tasks", selector, count)
-    return {"count": count}
-
-
-@rpc.method("Task.retry_from")
-async def task_retry_from(selector: str):
-    targets = await _select_tasks(selector)
-    from peagen.handlers import control_handler
-
-    count = await control_handler.apply(
-        "retry_from", queue, targets, READY_QUEUE, TASK_TTL
-    )
-    log.info("retry_from %s -> %d tasks", selector, count)
-    return {"count": count}
-
-
-@rpc.method("Guard.set")
-async def guard_set(label: str, spec: dict):
-    await queue.hset(f"guard:{label}", mapping=spec)
-    log.info("guard set %s", label)
-    return {"ok": True}
-
-
-@rpc.method("Task.patch")
-async def task_patch(taskId: str, changes: dict) -> dict:
-    """Update persisted metadata for an existing task."""
-    task = await _load_task(taskId)
-    if not task:
-        raise TaskNotFoundError(taskId)
-
-    for field, value in changes.items():
-        if field not in Task.model_fields:
-            continue
-        if field == "status":
-            value = Status(value)
-        setattr(task, field, value)
-
-    await _save_task(task)
-    await _persist(task)
-    await _publish_task(task)
-    if "result" in changes and isinstance(changes["result"], dict):
-        children = changes["result"].get("children")
-        if children:
-            for cid in children:
-                await _finalize_parent_tasks(cid)
-    log.info("task %s patched with %s", taskId, ",".join(changes.keys()))
-    return task.model_dump()
-
-
-@rpc.method("Task.get")
-async def task_get(taskId: str):
-    try:
-        uuid.UUID(taskId)
-    except ValueError:
-        raise RPCException(code=-32602, message="Invalid task id")
-    # hot cache
-    if t := await _load_task(taskId):
-        data = t.model_dump()
-        if t.duration is not None:
-            data["duration"] = t.duration
-        return data
-
-    # authoritative fallback (Postgres)
-    try:
-        return await get_task_result(taskId)  # raises TaskNotFoundError if missing
-    except TaskNotFoundError as exc:
-        # surface a proper JSON-RPC error so the envelope is valid
-        raise RPCException(code=ErrorCode.TASK_NOT_FOUND, message=str(exc))
-
-
-@rpc.method("Pool.listTasks")
-async def pool_list(poolName: str, limit: int | None = None, offset: int = 0):
-    """Return tasks queued for *poolName* with optional pagination."""
-
-    start = max(offset, 0)
-    end = -1 if limit is None else start + limit - 1
-    ids = await queue.lrange(f"{READY_QUEUE}:{poolName}", start, end)
-    tasks = []
-    for r in ids:
-        t = TaskRead.model_validate_json(r)
-        data = t.model_dump()
-        if t.duration is not None:
-            data["duration"] = t.duration
-        tasks.append(data)
-    return tasks
-
-
-# ─────────────────────────── Worker RPCs ────────────────────────
-@rpc.method("Worker.register")
-async def worker_register(
-    workerId: str,
-    pool: str,
-    url: str,
-    advertises: dict,
-    handlers: list[str] | None = None,
-):
-    """Register a worker and persist its advertised handlers."""
-
-    handler_list: list[str] = handlers or []
-    if not handler_list:
-        well_known_url = url.replace("/rpc", "/well-known")
-        try:
-            async with httpx.AsyncClient(timeout=5) as client:
-                resp = await client.get(well_known_url)
-                if resp.status_code == 200:
-                    handler_list = resp.json().get("handlers", [])
-        except Exception as exc:  # noqa: BLE001
-            log.warning("/well-known fetch failed for %s: %s", workerId, exc)
-
-    if not handler_list:
-        raise RPCException(code=-32602, message="worker supports no handlers")
-
-    await _upsert_worker(
-        workerId,
-        {
-            "pool": pool,
-            "url": url,
-            "advertises": advertises,
-            "handlers": handler_list,
-        },
-    )
-    log.info("worker %s registered (%s) handlers=%s", workerId, pool, handler_list)
-    return {"ok": True}
-
-
-@rpc.method("Worker.heartbeat")
-async def worker_heartbeat(
-    workerId: str, metrics: dict, pool: str | None = None, url: str | None = None
-):
-    # If gateway has no record (after crash), pool & url are expected
-    known = await queue.exists(WORKER_KEY.format(workerId))
-    if not known and not (pool and url):
-        log.warning(
-            "heartbeat from %s ignored: gateway lacks metadata; send pool+url or re-register",
-            workerId,
-        )
-        return {"ok": False}
-
-    await _upsert_worker(workerId, {"pool": pool, "url": url})
-    return {"ok": True}
-
-
-@rpc.method("Worker.list")
-async def worker_list(pool: str | None = None) -> list[dict]:
-    """Return active workers, optionally filtered by *pool*."""
-
-    keys = await queue.keys("worker:*")
-    workers = []
-    now = int(time.time())
-    for key in keys:
-        w = await queue.hgetall(key)
-        if not w:
-            continue
-        if now - int(w.get("last_seen", 0)) > WORKER_TTL:
-            continue
-        if pool and w.get("pool") != pool:
-            continue
-        workers.append(
-            {
-                "id": key.split(":", 1)[1],
-                **{k: v for k, v in w.items()},
-            }
-        )
-    return workers
-
-
-@rpc.method("Work.finished")
-async def work_finished(taskId: str, status: str, result: dict | None = None):
-    t = await _load_task(taskId)
-    if not t:
-        log.warning("Work.finished for unknown task %s", taskId)
-        return {"ok": False}
-
-    # update in-memory object
-    t.status = Status(status)
-    t.result = result
-    now = time.time()
-    if status == "running" and t.started_at is None:
-        t.started_at = now
-    elif Status.is_terminal(status):
-        if t.started_at is None:
-            t.started_at = now
-        t.finished_at = now
-
-    # persist everywhere
-    await _save_task(t)
-    await _persist(t)
-    await _publish_task(t)
-    if Status.is_terminal(status):
-        await _finalize_parent_tasks(taskId)
-
-    log.info("task %s completed: %s", taskId, status)
-    return {"ok": True}
 
 
 # ────────────────────────── Helpers ──────────────────────────
